basic maths project commit
This is a simple math library that provides basic functions such as addition, subtraction, multiplication, and division. It is packaged as a JAR file and can be easily used as a dependency in other Java projects.
I have also used this library in one of my projects, which I will include alongside it.